refactor(transport): Split transport/mod.rs into focused modules
Addresses issue #166 - Refactor transport module

Changes:
- Split transport/mod.rs (260 lines) into 6 focused files:
  * types.rs (62 lines): TransportInfo, TransportType,
    TransportCapabilities, FramingMethod
  * transport_trait.rs (42 lines): Transport trait definition
  * error.rs (55 lines): TransportError with conversions
  * config.rs (27 lines): TransportConfig with Default impl
  * factory.rs (39 lines): TransportFactory implementation
  * mod.rs (87 lines): Re-exports and tests
- Preserved existing submodules: connection, dynamic, http,
  stdio, websocket

Testing:
- All 377 tests passing
- No Clippy warnings
- Public API unchanged (no breaking changes)
